Refactored multiple modules to simplify logic, remove redundancies, and improve code readability.

+++ src/skills/avi_script/module/rand.rs
@@ -50,7 +50,7 @@ pub mod rand_functions {
             )
             .into())
         } else {
-            Ok(rand::rng().random_bool(probability as f64))
+            Ok(rand::rng().random_bool(probability))
         }
     }
